+++ HOTSPOT/src/main/webapp/WEB-INF/views/admin/inquiry/inquiryList.jsp
@@ -97,7 +97,14 @@
  	            		$("#check:checked").each(function(){		
  	            			selected.push($(this).val())
  	            		})
+ 	            		if(selected == ""){
+            				alert("선택된 항목이 없습니다.");
+            			}else{
+ 	            			if(confirm("삭제하시겠습니까?")){
+ 	            				
  	            			location.href="deleteList.inq?ino="+selected
+ 	            			}
+            			}
             		});
             	})
             </script>
